Refactor: 비동기 액션함수 분리
